feat:- customer profile templates added and customer details shown user profile page

+++ echobyte/order/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render , redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .models import Cart,CartItems 
 from product.models import Product, ProductVariant, ProductImage
 from .models import Cart,CartItems
@@ -6,10 +7,11 @@ from decimal import Decimal
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import JsonResponse
 from django.contrib import messages
+from django.urls import reverse
 
 
 # Create your views here.
-
+@login_required(login_url='signin')
 def cart(request):
     user = request.user
     try:
@@ -20,39 +22,48 @@ def cart(request):
         context = {'cart_items': None} 
     return render(request, 'cart.html', context)
 
+@login_required(login_url='signin')
 def add_to_cart(request):
     if request.method == 'POST':
         user = request.user
         customer = user
-        quantity = int(request.POST.get('quantity',1) )
+        quantity = int(request.POST.get('quantity', 1))
         product_id = int(request.POST.get('product-id'))
         product = ProductVariant.objects.get(pk=product_id)
         cart_obj, created = Cart.objects.get_or_create(owner=customer)
-        cart_item = CartItems.objects.create(product=product, cart=cart_obj, quantity=quantity,)
+        cart_item = CartItems.objects.create(product=product, cart=cart_obj, quantity=quantity)
         success_message = 'Product added to cart successfully.'
         return JsonResponse({'success_message': success_message})
     else:
-        return JsonResponse({'error': 'Invalid request'}, status=400)
+        # If it's not a POST request, return a JSON response indicating the need to sign in
+        return JsonResponse({'redirect_url': reverse('signin')})
+@login_required(login_url='signin')
 def delete_cart_item(request,pk):
-    cart_item = get_object_or_404(CartItems, pk=pk)
-    cart_item.delete()
-    return redirect('cart')
+    if request.user.is_authenticated:
+        cart_item = get_object_or_404(CartItems, pk=pk)
+        cart_item.delete()
+        return redirect('cart')
+    
+@login_required(login_url='signin')
 def add_cart_item_quantity(request,pk):
-    cart_item = CartItems.objects.get(pk=pk)
-    if cart_item.quantity >= 4:
-        messages.error(request, "Only 4 item can buy one order")
-    else: 
-        cart_item.quantity += 1
-        cart_item.save()
-    return redirect('cart')
+    if request.user.is_authenticated:
+        cart_item = CartItems.objects.get(pk=pk)
+        if cart_item.quantity >= 4:
+            messages.error(request, "Only 4 item can buy one order")
+        else: 
+            cart_item.quantity += 1
+            cart_item.save()
+        return redirect('cart')
+@login_required(login_url='signin')
 def sub_cart_item_quantity(request, pk):
-    cart_item = CartItems.objects.get(pk=pk)
-    if cart_item.quantity <= 1:
-        messages.error(request, "At least one item is needed in the order.")
-    else:
-        cart_item.quantity -= 1
-        cart_item.save()
-    return redirect('cart')
+    if request.user.is_authenticated:
+        cart_item = CartItems.objects.get(pk=pk)
+        if cart_item.quantity <= 1:
+            messages.error(request, "At least one item is needed in the order.")
+        else:
+            cart_item.quantity -= 1
+            cart_item.save()
+        return redirect('cart')
